+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### Fixed ###
 
+  * Bags would not always close when hitting the Escape key
   * [Various bugs and errors with the UI skin.](https://www.wowinterface.com/downloads/info18589-Aurora.html#changelog)