+++ CHANGELOG.md
@@ -12,7 +12,8 @@
 
 ### Changed ###
 
-  * The "New Items" filter bag will now retain it items until the reset button is clicked
+  * Tweaked xp bar colors when rested
+  * The "New Items" filter bag will now retain items until it's explicitly reset or a UI reload
   * The default positions of the zone ability and extra action have been updated
 
 ### Fixed ###