+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 
 ### Fixed ###
 
+  * Partial Soul Shards were not shown for Destro Locks
   * Error when mousing over characters with low level artifact weapons
   * The UI scale button was in the wrong place
   * Bags would not always close when hitting the Escape key