+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [Unreleased] ##
+﻿## [2.2.6] ##
 ### Modified AddOns ###
 
   * RealUI_Skins
@@ -388,6 +388,7 @@
 
 
 [Unreleased]: https://github.com/RealUI/RealUI/compare/master...develop
+[2.2.6]: https://github.com/RealUI/RealUI/compare/2.2.5...2.2.6
 [2.2.5]: https://github.com/RealUI/RealUI/compare/2.2.4...2.2.5
 [2.2.4]: https://github.com/RealUI/RealUI/compare/2.2.3...2.2.4
 [2.2.3]: https://github.com/RealUI/RealUI/compare/2.2.2...2.2.3